Split PlantingSiteStoreTest
`PlantingSiteStoreTest` was getting pretty large, and the upcoming changes to
support planting site map editing would have made it even larger. Split it up
into smaller classes.

No changes to any of the actual test code here, just moving the test methods
to new classes.
