+++ src/test/kotlin/com/terraformation/backend/tracking/db/plantingSiteStore/PlantingSiteStoreUpdateSiteTest.kt
@@ -1,0 +1,152 @@
+package com.terraformation.backend.tracking.db.plantingSiteStore
+
+import com.terraformation.backend.customer.event.PlantingSiteTimeZoneChangedEvent
+import com.terraformation.backend.db.ProjectInDifferentOrganizationException
+import com.terraformation.backend.db.default_schema.OrganizationId
+import com.terraformation.backend.db.tracking.tables.pojos.PlantingSiteHistoriesRow
+import com.terraformation.backend.db.tracking.tables.pojos.PlantingSitesRow
+import com.terraformation.backend.multiPolygon
+import com.terraformation.backend.point
+import com.terraformation.backend.tracking.model.PlantingSiteDepth
+import com.terraformation.backend.tracking.model.PlantingSiteModel
+import com.terraformation.backend.util.Turtle
+import io.mockk.every
+import java.math.BigDecimal
+import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
+import org.springframework.security.access.AccessDeniedException
+
+internal class PlantingSiteStoreUpdateSiteTest : PlantingSiteStoreTest() {
+  @Nested
+  inner class UpdatePlantingSite {
+    @Test
+    fun `updates values`() {
+      val initialModel =
+          store.createPlantingSite(
+              PlantingSiteModel.create(
+                  boundary = Turtle(point(0)).makeMultiPolygon { square(100) },
+                  name = "initial name",
+                  organizationId = organizationId,
+                  timeZone = timeZone,
+              ))
+
+      val createdTime = clock.instant()
+      val newTimeZone = insertTimeZone("Europe/Paris")
+      val now = createdTime.plusSeconds(1000)
+      clock.instant = now
+
+      val newBoundary = Turtle(point(1)).makeMultiPolygon { square(200) }
+
+      store.updatePlantingSite(initialModel.id, emptyList()) { model ->
+        model.copy(
+            boundary = newBoundary,
+            description = "new description",
+            name = "new name",
+            timeZone = newTimeZone,
+        )
+      }
+
+      assertEquals(
+          listOf(
+              PlantingSitesRow(
+                  areaHa = BigDecimal("4.0"),
+                  boundary = newBoundary,
+                  gridOrigin = initialModel.gridOrigin,
+                  id = initialModel.id,
+                  organizationId = organizationId,
+                  name = "new name",
+                  description = "new description",
+                  createdBy = user.userId,
+                  createdTime = createdTime,
+                  modifiedBy = user.userId,
+                  modifiedTime = now,
+                  timeZone = newTimeZone,
+              )),
+          plantingSitesDao.findAll(),
+          "Planting sites")
+
+      assertEquals(
+          setOf(
+              PlantingSiteHistoriesRow(
+                  plantingSiteId = initialModel.id,
+                  createdBy = user.userId,
+                  createdTime = createdTime,
+                  boundary = initialModel.boundary,
+                  gridOrigin = initialModel.gridOrigin,
+              ),
+              PlantingSiteHistoriesRow(
+                  plantingSiteId = initialModel.id,
+                  createdBy = user.userId,
+                  createdTime = clock.instant,
+                  boundary = newBoundary,
+                  gridOrigin = initialModel.gridOrigin,
+              ),
+          ),
+          plantingSiteHistoriesDao.findAll().map { it.copy(id = null) }.toSet(),
+          "Planting site histories")
+    }
+
+    @Test
+    fun `ignores boundary updates on detailed planting sites`() {
+      val plantingSiteId = insertPlantingSite(boundary = multiPolygon(1))
+      insertPlantingZone()
+
+      store.updatePlantingSite(plantingSiteId, emptyList()) { model ->
+        model.copy(boundary = multiPolygon(2))
+      }
+
+      assertEquals(multiPolygon(1), plantingSitesDao.findAll().first().boundary)
+    }
+
+    @Test
+    fun `publishes event if time zone updated`() {
+      val plantingSiteId = insertPlantingSite(timeZone = timeZone)
+      val initialModel = store.fetchSiteById(plantingSiteId, PlantingSiteDepth.Site)
+      val newTimeZone = insertTimeZone("Europe/Paris")
+
+      store.updatePlantingSite(plantingSiteId, emptyList()) { it.copy(timeZone = newTimeZone) }
+
+      val expectedEvent =
+          PlantingSiteTimeZoneChangedEvent(
+              initialModel.copy(timeZone = newTimeZone), timeZone, newTimeZone)
+
+      eventPublisher.assertEventPublished(expectedEvent)
+    }
+
+    @Test
+    fun `does not publish event if time zone not updated`() {
+      val plantingSiteId = insertPlantingSite(timeZone = timeZone)
+
+      store.updatePlantingSite(plantingSiteId, emptyList()) { it.copy(description = "edited") }
+
+      eventPublisher.assertEventNotPublished(PlantingSiteTimeZoneChangedEvent::class.java)
+    }
+
+    @Test
+    fun `throws exception if no permission`() {
+      val plantingSiteId = insertPlantingSite()
+
+      every { user.canUpdatePlantingSite(any()) } returns false
+
+      assertThrows<AccessDeniedException> {
+        store.updatePlantingSite(plantingSiteId, emptyList()) { it }
+      }
+    }
+
+    @Test
+    fun `throws exception if project is in a different organization`() {
+      val plantingSiteId = insertPlantingSite()
+      val otherOrganizationId = OrganizationId(2)
+      insertOrganization(otherOrganizationId)
+      val otherOrgProjectId = insertProject(organizationId = otherOrganizationId)
+
+      assertThrows<ProjectInDifferentOrganizationException> {
+        store.updatePlantingSite(plantingSiteId, emptyList()) {
+          it.copy(projectId = otherOrgProjectId)
+        }
+      }
+    }
+  }
+}
